feat(mobile): plan cards in onboarding, replay from Settings
The plan step now shows the paywall's own tier cards, loaded from the same
catalog and priced by StoreKit. TierCard and SpecRow move out of plan.tsx into
src/omg/tier-card.tsx so both screens render a tier in one shape. A card tap
opens the paywall; "Start free" stays.

Settings gains "Replay the welcome tour", a plain route (app/onboarding.tsx)
that runs the intro then setup without touching the first-run flags. The
intro's last button reads "Continue" there instead of "Sign in".

rosterFromReadiness() is the one derivation of the connect step's input,
used by the first-run gate and by replay.

Workflow: a `simulator` profile option, so a dev client with new native
modules can be built on a runner and looked at.
